@@ -2,7 +2,7 @@ name: instagram_clone
 description: A new Flutter project.
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.19.3 <3.0.0'
+  sdk: ">=3.1.3 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -28,25 +28,25 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
-  cached_network_image: ^3.0.0
-  cloud_firestore: ^4.7.1
-  cupertino_icons: ^1.0.2
-  firebase_analytics: ^10.4.5
-  firebase_auth: ^4.6.1
-  firebase_core: ^2.13.0
-  firebase_messaging: ^14.6.7
-  firebase_storage: ^11.2.1
+  cached_network_image:
+  cloud_firestore:
+  cupertino_icons:
+  firebase_analytics:
+  firebase_auth:
+  firebase_core:
+  firebase_messaging: ^15.1.3
+  firebase_storage:
   flutter:
     sdk: flutter
-  flutter_local_notifications: ^15.1.1
-  flutter_staggered_grid_view: ^0.6.2
+  flutter_local_notifications: ^17.2.3
+  flutter_staggered_grid_view:
   flutter_svg: null
-  image_picker: ^0.8.7+5
-  intl: ^0.18.1
-  provider: ^6.0.5
-  shimmer: ^2.0.0
-  story_view: ^0.15.0
-  uuid: ^3.0.7
+  image_picker:
+  intl:
+  provider:
+  shimmer:
+  story_view: ^0.16.5
+  uuid:
   # flutterfire_cli: ^0.2.7
 
 dev_dependencies:
@@ -64,7 +64,6 @@ flutter_icons:
 # following page: https://dart.dev/tools/pub/pubspec
 # The following section is specific to Flutter packages.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
